Build ONNX Server application on Windows (#57)
* Gets Boost compiling on Windows

* Fix integer conversion and comparison problems

* Use size_t in converter_tests instead of int

* Fix hosting integration tests on Windows

* Removes checks for port because it's an unsigned short

* Fixes comparison between signed and unsigned data types

* Pip install protobuf and numpy

+++ onnxruntime/server/http/predict_request_handler.cc
@@ -112,7 +112,7 @@ static bool ParseRequestPayload(const HttpContext& context, SupportedContentType
       break;
     }
     case SupportedContentType::PbByteArray: {
-      bool parse_succeeded = predictRequest.ParseFromArray(body.data(), body.size());
+      bool parse_succeeded = predictRequest.ParseFromArray(body.data(), static_cast<int>(body.size()));
       if (!parse_succeeded) {
         error_code = http::status::bad_request;
         error_message = "Invalid payload.";
